admin_site/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3
